+++ packages/core/src/components/button/button.stories.tsx
@@ -19,7 +19,6 @@ export default {
         stretch: { control: 'boolean' },
     },
 } as Meta<typeof Button>;
-
 type Story = StoryObj<typeof Button>;
 
 export const Default: Story = {
@@ -27,7 +26,7 @@ export const Default: Story = {
         <HStack gap="$200">
             <Button {...args}>Button</Button>
             <Button {...args} asChild>
-                <a href="https://vapor.goorm.io">Link Button(Polymorphic)</a>
+                <a href="https://vapor-ui.goorm.io">Link Button(Polymorphic)</a>
             </Button>
         </HStack>
     ),
@@ -65,7 +64,7 @@ export const TestBed: Story = {
 const Buttons = ({ color, disabled }: ButtonProps) => {
     return (
         <div style={{ display: 'flex', flexDirection: 'column' }}>
-            <Button color={color} disabled={disabled} variant="fill" gap="050">
+            <Button color={color} disabled={disabled} variant="fill">
                 {upperFirst(`${color}`)}
             </Button>
             <Button color={color} disabled={disabled} variant="ghost">