+++ packages/core/src/components/checkbox/checkbox.stories.tsx
@@ -38,7 +38,7 @@ export const Default: Story = {
         };
 
         return (
-            <Flex flexDirection="column">
+            <Flex style={{ flexDirection: 'column' }}>
                 <Text typography="heading3">Uncontrolled</Text>
                 <Checkbox.Root {...args}>
                     <Checkbox.Control />
@@ -79,8 +79,8 @@ export const Default: Story = {
 export const TestBed: Story = {
     render: () => {
         return (
-            <Flex flexDirection="column" gap="$100">
-                <Flex gap="$150">
+            <Flex style={{ gap: 'var(--vapor-size-dimension-100)', flexDirection: 'column' }}>
+                <Flex style={{ gap: 'var(--vapor-size-dimension-150)' }}>
                     <Checkbox.Root>
                         <Checkbox.Label>Default</Checkbox.Label>
                         <Checkbox.Control />
@@ -95,7 +95,7 @@ export const TestBed: Story = {
                     </Checkbox.Root>
                 </Flex>
 
-                <Flex gap="$150">
+                <Flex style={{ gap: 'var(--vapor-size-dimension-150)' }}>
                     <Checkbox.Root disabled>
                         <Checkbox.Label>Disabled</Checkbox.Label>
                         <Checkbox.Control />
@@ -110,7 +110,7 @@ export const TestBed: Story = {
                     </Checkbox.Root>
                 </Flex>
 
-                <Flex gap="$150">
+                <Flex style={{ gap: 'var(--vapor-size-dimension-150)' }}>
                     <Checkbox.Root invalid>
                         <Checkbox.Label>Invalid</Checkbox.Label>
                         <Checkbox.Control />