+++ packages/core/src/components/grid/grid.tsx
@@ -1,42 +1,41 @@
 'use client';
 
-import type { CSSProperties } from 'react';
+import type { CSSProperties, ComponentPropsWithoutRef } from 'react';
 import { forwardRef } from 'react';
 
 import { assignInlineVars } from '@vanilla-extract/dynamic';
 import clsx from 'clsx';
 
-import type { VaporComponentProps } from '~/libs/factory';
-import { vapor } from '~/libs/factory';
 import { createSplitProps } from '~/utils/create-split-props';
 
-import type { GridRootVariants } from './grid.css';
+import { Box } from '../box';
+import type { RootVariants } from './grid.css';
 import * as styles from './grid.css';
 
 /* -------------------------------------------------------------------------------------------------
  * Grid
  * -----------------------------------------------------------------------------------------------*/
 
-type GridPrimitiveProps = VaporComponentProps<'div'>;
-type GridVariants = GridRootVariants & {
+type GridPrimitiveProps = ComponentPropsWithoutRef<typeof Box>;
+type GridVariants = RootVariants & {
     inline?: boolean;
     templateRows?: string;
     templateColumns?: string;
     flow?: CSSProperties['gridAutoFlow'];
 };
 
-type GridRootProps = GridPrimitiveProps & GridVariants;
+interface GridRootProps extends GridPrimitiveProps, GridVariants {}
 
 const Root = forwardRef<HTMLDivElement, GridRootProps>(
     ({ className, style, children, ...props }, ref) => {
-        const [gridRootProps, otherProps] = createSplitProps<GridVariants>()(props, [
+        const [variantProps, otherProps] = createSplitProps<GridVariants>()(props, [
             'inline',
             'templateRows',
             'templateColumns',
             'flow',
         ]);
 
-        const { inline, templateRows, templateColumns, ...variants } = gridRootProps;
+        const { inline, templateRows, templateColumns, ...variants } = variantProps;
 
         const cssVariables = assignInlineVars({
             [styles.gridTemplateRows]: templateRows,
@@ -44,7 +43,7 @@ const Root = forwardRef<HTMLDivElement, GridRootProps>(
         });
 
         return (
-            <vapor.div
+            <Box
                 ref={ref}
                 display={inline ? 'inline-grid' : 'grid'}
                 style={{ ...cssVariables, ...style }}
@@ -52,7 +51,7 @@ const Root = forwardRef<HTMLDivElement, GridRootProps>(
                 {...otherProps}
             >
                 {children}
-            </vapor.div>
+            </Box>
         );
     },
 );
@@ -62,10 +61,10 @@ Root.displayName = 'Grid';
  * Grid.Item
  * -----------------------------------------------------------------------------------------------*/
 
-type GridItemPrimitiveProps = VaporComponentProps<'div'>;
+type GridItemPrimitiveProps = ComponentPropsWithoutRef<typeof Box>;
 type GridItemVariants = { rowSpan?: string; colSpan?: string };
 
-type GridItemProps = GridItemPrimitiveProps & GridItemVariants;
+interface GridItemProps extends GridItemPrimitiveProps, GridItemVariants {}
 
 const Item = forwardRef<HTMLDivElement, GridItemProps>(
     ({ rowSpan, colSpan, className, style, children, ...props }, ref) => {
@@ -75,14 +74,14 @@ const Item = forwardRef<HTMLDivElement, GridItemProps>(
         });
 
         return (
-            <vapor.div
+            <Box
                 ref={ref}
                 style={{ ...cssVariables, ...style }}
                 className={clsx(styles.item, className)}
                 {...props}
             >
                 {children}
-            </vapor.div>
+            </Box>
         );
     },
 );