+++ apps/website/CHANGELOG.md
@@ -1,5 +1,22 @@
 # website
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [e139a7f]
+- Updated dependencies [03655b1]
+- Updated dependencies [488c614]
+    - @vapor-ui/core@0.3.0
+
+## 0.0.3
+
+### Patch Changes
+
+- 6749d80: fix(createThemeConfig): support RSC
+- Updated dependencies [6749d80]
+    - @vapor-ui/core@0.2.1
+
 ## 0.0.2
 
 ### Patch Changes